pyproject.toml
[project]
name = "pixiv-mcp-server"
version = "2.0.0"
description = "A powerful Pixiv toolset for Large Language Models via MCP."
authors = [{ name = "AI", email = "ai@example.com" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.0.0",
    "pixivpy3>=3.7.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/222wcnm/pixiv-mcp-server"
Issues = "https://github.com/222wcnm/pixiv-mcp-server/issues"

[project.scripts]
pixiv-mcp-server = "pixiv_mcp_server.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pixiv_mcp_server*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
